Fix #62: Allow arguments with '-' in their name
Nubia expects you to use '_' in argument names to match up with the
characters python variable names allow. Nubia automatically converts
'_' to '-' in argument name when showing the possible arguments. Thus,
'start_time' is shown as 'start-time'. However, when attempting to
match up the supplied arguments with what's supported, it fails to
convert the '-' to '_' and so the match against supported arguments
fails.

This fix ensures that the check against expected arguments is made by
first converting the '-' in supplied arguments to '_'.