Fix #62: Allow arguments with '-' in their name (#63)
Summary:
Nubia expects you to use '_' in argument names to match up with the
characters python variable names allow. Nubia automatically converts
'_' to '-' in argument name when showing the possible arguments. Thus,
'start_time' is shown as 'start-time'. However, when attempting to
match up the supplied arguments with what's supported, it fails to
convert the '-' to '_' and so the match against supported arguments
fails.

This fix ensures that the check against expected arguments is made by
first converting the '-' in supplied arguments to '_'.
Pull Request resolved: https://github.com/facebookincubator/python-nubia/pull/63

Reviewed By: jpdborgna

Differential Revision: D20337923

Pulled By: AhmedSoliman

fbshipit-source-id: 928e9db6bdbb2988b070a5ebf29103d099c0d01b